requirements.txt
streamlit>=1.30.0
python-docx>=0.8.11
openpyxl>=3.1.2
